pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.75.0",
    "bcrypt>=5.0.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
]
